+++ curated_transformers/generation/default_generator.py
@@ -5,9 +5,9 @@ import torch
 
 from ..models.module import CausalLMModule
 from ..quantization import BitsAndBytesConfig
-from ..tokenization.auto_tokenizer import AutoTokenizer
-from ..tokenization.chunks import InputChunks, TextChunk
-from ..tokenization.tokenizer import TokenizerBase
+from ..tokenizers.auto_tokenizer import AutoTokenizer
+from ..tokenizers.chunks import InputChunks, TextChunk
+from ..tokenizers.tokenizer import TokenizerBase
 from ..util.auto_model import AutoCausalLM
 from .config import GeneratorConfig, SampleGeneratorConfig
 from .generator import Generator
