+++ curated_transformers/tests/tokenization/test_llama_tokenizer.py
@@ -1,7 +1,7 @@
 import pytest
 
 from curated_transformers._compat import has_hf_transformers
-from curated_transformers.tokenization import LLaMATokenizer
+from curated_transformers.tokenizers import LLaMATokenizer
 
 from .util import compare_tokenizer_outputs_with_hf_tokenizer
