+++ curated_transformers/tests/tokenization/test_roberta_tokenizer.py
@@ -2,7 +2,7 @@ import pytest
 import torch
 
 from curated_transformers._compat import has_hf_transformers
-from curated_transformers.tokenization import PiecesWithIds, RobertaTokenizer
+from curated_transformers.tokenizers import PiecesWithIds, RobertaTokenizer
 
 from ..util import torch_assertclose
 from .util import compare_tokenizer_outputs_with_hf_tokenizer
