+++ docs/source/tokenizers.rst
@@ -15,14 +15,14 @@ that should be tokenized. The other type of supported chunk is the
 :class:`.SpecialPieceChunk`. The piece stored by this type of chunk is not
 tokenized but looked up directly.
 
-.. autoclass:: curated_transformers.tokenization.InputChunks
+.. autoclass:: curated_transformers.tokenizers.InputChunks
    :members:
    :show-inheritance:
 
-.. autoclass:: curated_transformers.tokenization.SpecialPieceChunk
+.. autoclass:: curated_transformers.tokenizers.SpecialPieceChunk
    :members:
 
-.. autoclass:: curated_transformers.tokenization.TextChunk
+.. autoclass:: curated_transformers.tokenizers.TextChunk
    :members:
 
 Pieces
@@ -31,7 +31,7 @@ Pieces
 All tokenizers decode raw strings into pieces. The pieces are
 stored in a special container :class:`.PiecesWithIds`.
 
-.. autoclass:: curated_transformers.tokenization.PiecesWithIds
+.. autoclass:: curated_transformers.tokenizers.PiecesWithIds
    :members:
    :show-inheritance:
 
@@ -44,65 +44,65 @@ committing to a specific tokenizer type, you can use the :class:`.AutoTokenizer`
 class. This class also provides a ``from_hf_hub`` method to load a tokenizer,
 but will try to infer the tokenizer type automatically.
 
-.. autoclass:: curated_transformers.tokenization.AutoTokenizer
+.. autoclass:: curated_transformers.tokenizers.AutoTokenizer
    :members:
 
 Tokenizers
 ----------
 
-.. autoclass:: curated_transformers.tokenization.Tokenizer
+.. autoclass:: curated_transformers.tokenizers.Tokenizer
    :members:
    :special-members: __call__
    :show-inheritance:
 
-.. autoclass:: curated_transformers.tokenization.TokenizerBase
+.. autoclass:: curated_transformers.tokenizers.TokenizerBase
    :members:
    :special-members: __call__
    :show-inheritance:
 
-.. autoclass:: curated_transformers.tokenization.ByteBPETokenizer
-   :members:
-   :inherited-members:
-   :special-members: __call__
-   :show-inheritance:
-
-.. autoclass:: curated_transformers.tokenization.WordPieceTokenizer
+.. autoclass:: curated_transformers.tokenizers.ByteBPETokenizer
    :members:
    :inherited-members:
    :special-members: __call__
    :show-inheritance:
 
-.. autoclass:: curated_transformers.tokenization.SentencePieceTokenizer
+.. autoclass:: curated_transformers.tokenizers.WordPieceTokenizer
    :members:
    :inherited-members:
    :special-members: __call__
    :show-inheritance:
 
-.. autoclass:: curated_transformers.tokenization.RobertaTokenizer
+.. autoclass:: curated_transformers.tokenizers.SentencePieceTokenizer
    :members:
    :inherited-members:
    :special-members: __call__
    :show-inheritance:
 
-.. autoclass:: curated_transformers.tokenization.BertTokenizer
+.. autoclass:: curated_transformers.tokenizers.RobertaTokenizer
    :members:
    :inherited-members:
    :special-members: __call__
    :show-inheritance:
 
-.. autoclass:: curated_transformers.tokenization.CamembertTokenizer
+.. autoclass:: curated_transformers.tokenizers.BertTokenizer
    :members:
    :inherited-members:
    :special-members: __call__
    :show-inheritance:
 
-.. autoclass:: curated_transformers.tokenization.LLaMATokenizer
+.. autoclass:: curated_transformers.tokenizers.CamembertTokenizer
    :members:
    :inherited-members:
    :special-members: __call__
    :show-inheritance:
 
-.. autoclass:: curated_transformers.tokenization.XlmrTokenizer
+.. autoclass:: curated_transformers.tokenizers.LLaMATokenizer
+   :members:
+   :inherited-members:
+   :special-members: __call__
+   :show-inheritance:
+
+.. autoclass:: curated_transformers.tokenizers.XlmrTokenizer
    :members:
    :inherited-members:
    :special-members: __call__
